+++ ShoesShop.Application/DTOs/Carts/UpdateCartItemRequest.cs
@@ -2,7 +2,7 @@ namespace ShoesShop.Application.DTOs.Carts;
 
 public class UpdateCartItemRequest
 {
-    public int CartItemId { get; set; }
+    public long CartItemId { get; set; }
 
     public int Quantity { get; set; }
 }
