+++ ShoesShop.Application/Interfaces/Services/ICartService.cs
@@ -4,13 +4,13 @@ namespace ShoesShop.Application.Interfaces.Services;
 
 public interface ICartService
 {
-    Task<CartResponse> AddToCartAsync(int userId, AddCartItemRequest request);
+    Task<CartResponse> AddToCartAsync(long userId, AddCartItemRequest request);
 
-    Task<CartResponse> GetCartAsync(int userId);
+    Task<CartResponse> GetCartAsync(long userId);
 
-    Task<CartResponse> UpdateQuantityAsync(int userId, UpdateCartItemRequest request);
+    Task<CartResponse> UpdateQuantityAsync(long userId, UpdateCartItemRequest request);
 
-    Task DeleteCartItemAsync(int userId, int cartItemId);
+    Task DeleteCartItemAsync(long userId, long cartItemId);
 
-    Task ClearCartAsync(int userId);
+    Task ClearCartAsync(long userId);
 }
